Support replacing logger; rename loggertest package to rslogtest
