+++ pkg/rslog/rslogtest/mocks.go
@@ -1,4 +1,4 @@
-package loggertest
+package rslogtest
 
 // Copyright (C) 2021 by RStudio, PBC.
